Lower the minimum pump duty to 50%
This has been an ongoing discussion in krakenx and it took me a long
time to decide what my stance is.

The main issue is that while allowed by the firmware, such low speeds
are not (currently) exposed in CAM.  This probably means that there are
scenarios where correlation has been observed between pump speeds lower
than 60% and increased failure rates.

However, no public statement has been made, and there should also be
plenty of scenarios where they are perfectly harmless, for the following
reasons:

 - they are allowed by the firmware
 - the pump would still run well above its minimum specified speed (i.e.
   1900+ rpm, when the specification lists 1600 ± 300 rpm)[1]
 - the performance differences between 1900 rpm (50%), 2100 rpm (60%) and
   even 2800 rpm (100%) are very small, both on liquid and component
   temperatures[2][3]
 - while DC controlled devices might stall at low voltages, these pumps
   are controlled with (high-frequency) PWM, thus greatly reducing the
   likelihood of this issue[4]

For these reasons, my position is (finally) that we should educate the
users, instead of limiting them.

[1] https://web.archive.org/web/20180918051402/https://www.nzxt.com/products/kraken-x62
[2] According to my own testing with Kraken X62 + i7 8700K + 170+ W
[3] https://youtu.be/3B7ou8fOFfE?t=1m35s
[4] https://youtu.be/Kc07JtDerUw