Tune down warnings about failure to pre open the device
Keith reported on issue #16 ("Need liquidctl to work with EVGA CLC 280
AIO") that pre open fails on the CLC 280 with

	usb.core.USBError: [Errno 32] Pipe error

This tunes down down the warnings, while at the same time making them
clearer and more informative.  We still want to know which devices don't
support these methods, and when.