Remove --dry-run as it was not functional or used
As far as we know, this feature was *never* used outside of the
krakencurve-poc.  Thus it will be removed immediately, without a
deprecation period.

Additionally, it had never been properly implemented,
and some device state *always* changed (e.g. `connect()` would set the
configuration, which can clear device registers.

Finally, a true implementation of `--dry-run` would make it useless, as
the driver would not be able to recognize device features and any
emulated behavior would be incredibly limited and far for reality.