Remove normalization and automatic optimization of fan profiles
The device already appears to correctly sanitize the profile, and also
already features interpolation between the supplied points.

By externally normalizing and trying to optimize the profile we were
removing a great deal of control from the user, and not really providing
anything useful in return.  Not to mention autofill_profile has a couple
of quirks.