Fix instantiating v1.1.0 era drivers directly with usb.core.Device
We'll retain compatibility for now because in v1.1.0 there was no cleaner
way to extend a driver.  However, this should not be depended on in new
code.(note 1)

Instead prefer to use find_supported_devices, which now allows and
forwards keyword arguments to the subclass constructor, and avoid
instantiating the device directly.

Note 1: the only known use of the legacy behavior is code written
myself, that will be updated as soon as v1.2.0 is released.  It also
only touches a single driver.