Add legacy mode for Asetek 690LC coolers and support 2nd gen Krakens
The legacy devices only support a very specific order of messages, and
do not support enough commands to allow the configuration of each
feature separately (see: jaksi/leviathan/PROTOCOL.md)

To bypass this limitation, the settings of previous invocations of the
driver APIs are stored in a auxiliary location, and are reloaded when
it is necessary to send messages beyond those specifically related to a
particular function call.

The idea was originally proposed and tested by Chris Griffith (see:
cdgriffith/liquidctl@0b6845e540c4).  The approach taken here diverges
from his original implementation in a few ways, but the basic idea of
transparently storing the data required for subsequent calls remains.

jaksi/leviathan/PROTOCOL.md:
		https://github.com/jaksi/leviathan/blob/master/PROTOCOL.md

cdgriffith/liquidctl@0b6845e540c4:
		https://github.com/cdgriffith/liquidctl/commit/0b6845e540c4