Remove serial number from settings path
It can consist of an arbitrary Unicode string, pottentially containing
forbidden characters or .. (reference to parent directory).

Since these files are for internal use and only expected to last as long
as the system is running, it is reasonable to identify them solely with
the  bus name and port numbers (besides vid_pid_release).