Add EVGA CLC 360 to the supported devices
Besides physical similarities, lack of new features and a compatible
model number, EVGA has not released a new version of Flow Control for
the new cooler, announced yesterday.

This strongly suggests that the CLC 360 is entirely compatible with
previous implementations and, most likely, combines the stablish ed
pump/block/PCB design with a larger radiator.