Pass alert threshold and color along even in blackout mode
While on the EVGA CLC tested 'blackout' disables everything, including
the high temperature alert, this might not always be the case on
different firmware versions or Asetek 690LC coolers.

If these parameters have been supplied we don't want to override them
with the defaults, in case they don't go unused.