Make sure 690LC are always released
This is done as soon as possible to allow allow other programs to
(attempt to) interact with the device when we're not using it.

However the cost of this approach be slightly higher on these devices
since we always release them after connect(), even if we're just about
to do something else with it.