Catch the USBError when there's no active configuration
Unlike what the PyUSB FAQ suggests, get_active_configuration raises an
error if the device is in unconfigured state.

Note that even after set_configuration the device is still not ready.
This is still an open FIXME, and for now the device will only be
functional on the *next* invocation of liquidctl.