Implement fan control for Seasonic E PSUs
Uses the FAN_COMMAND_1 (0x3b) PMBus command, with the desired duty cycle
encoded in LINEAR11, and a PEC byte.

---

In the captures supplied by Jon[1][2] we can't see any missing commands
or bugs that would explain why fan control wasn't working in the initial
implementation.  There is however an extra trailing byte passed with
FAN_COMMAND_1, and it clearly isn't leftover data from a previous use of
the buffer.  Instead, it could be a Packed Error Code (PEC).

			     ac04603b17009b
			     ac04603b0000a7
				         ^^

Unfortunately confirming this was not trivial and all the first attempts
of validating the observed PECs failed.

After that a lot of time was spent looking at captures, reading the
specs _again_ and verifying the CRC implementation, until...

The PMBus spec states that the PEC must be computed for the entire
message.  It also states that the message always starts with an address
byte, which is then followed by the command byte.

So, the first thing we can do is compute the PEC only from the byte that
precedes the command byte forward.  Additionally, Jon had already
speculated early on[3] that 0x60 could be the address.

			     ac04603b17009b
				 ^^^^^^^^

While this appeared to make sense, it still didn't work.

However, I had previously disagreed with him on 0x60 being an address
because, according to the PMBus spec, the address byte should contain
the address on the 7 most significant bits and a read (1) or write (0)
least significant bit.  Essentially, we should expect to see 0x60 only on
writes, such as these executions of FAN_COMMAND_1, but on all reads we
should see 0x61, which we don't.

Where we do see a pattern of LSB 1 for read and 0 for write is in the
first byte of all messages sent to the device.  But we now know that the
other bits in that first byte can't be the PMBus slave address, because
of how the PEC has to be computed and later validated by the slave
itself.

So what if 0x60 is actually the slave *address*, not the address byte?

Well, shifting it left 1 bit seems to confirm that this is the case.
At least we get a successful PEC check when we interpret it like that!

So, in the hope that his will allow fan control to work, this patch adds
a PEC to the FAN_COMMAND_1 writes.

[1]
https://github.com/jonasmalacofilho/liquidctl-device-data/tree/master/NZXT%20E500/01%20-%20generic%20capture%20-%20jnettlet
[2]
https://github.com/jonasmalacofilho/liquidctl-device-data/tree/master/NZXT%20E500/02%20-%20generic%20capture%20-%20jnettlet
[3]
https://github.com/jonasmalacofilho/liquidctl/issues/31#issuecomment-500101710

---

Comparing the captures against the debug data supplied by Ivan shows
that, apparently, most of the response from a write word command is
simply whatever was left on the device output buffer.[4]

Thus, it seems that the only check we can do is for a first byte ==
0xaa.

[4] https://github.com/jonasmalacofilho/liquidctl/pull/55#issuecomment-550410701