Revise status command to display Unknowns
Because we don't yet know the codes for 300mm Underglow and 200/250mm LED strips, and HUE 2 Ambient, those devices will now appear as 'Unknown'. The user should rerun the command in --debug mode to see the full reply from the device. If they provide this reply, we can update the code.