Default LED color changed to black from white
For color modes where all 40 LEDs can be individually specified, the remaining LEDs are not set to black (off) if no color is specified. For example, use can specify colors for the first 8 or 10 LEDs, in which case all remaining LEDs will be turned off. We may change this behavior in the future to allow user to specify a default color for the remaining LEDs.