Set minimum required Python version to 3.6
Up to version 1.2.0 liquidctl would probably run on Python 3.5 or even
3.4, but it was only tested on 3.6 or later.

One big issue is that we can't _really_ test the code with access to the
hardware, and all involved in the repository (even issue authors) seem
to run modern versions only.  Python 3.6 is only throughly tested
because it's what we ship in the pre-built binaries for Windows.

Adopting Python 3.6 as a minimum requirement not only matches what's
actually tested, but also allows us to use newer features (e.g.
f-strings).  We also use IntFlag in the PMBus code for the new drivers.

---

If there's enough demand (and thus testing) for support on older Python
versions, we can revisit this.