Increase resolution of kraken_two fan and pump profiles
Due to a firmware limitation the same set of temperatures must be
used on both channels.  This persists even on recent firmware versions
(tested with 6.0.2).

However, the number of temperature–duty points is not limited to 21 as
previously thought.  Up to 61 points appear to work well, and the base
index offset between channels suggests that up to 63 should be possible.

Still, we prefer to keep the number of writes low by trimming the
interval and resolution to the most useful ranges.  Effectively, we
have kept the same range (20°–60°C) as before, with increased (optimal)
resolution bellow 50°C.