Add support for HSV, HSL and explicit RGB representations of colors
When configuring lighting animations it is sometimes useful to be able
to do color transformations directly on the shell.

An old-school (yet still used today in many GUIs) method to simple
colors transformations is to use alternative representations of the RGB
color model, where hue, relative chroma (saturation) and value/lightness
can be directly manipulated.

Enhance color_from_str to be able to parse and translate such alternate
representations.  In addition to the HSV and HSL models, also add an
explicit/decimal model for the standard RGB representation.

This results in the following supported formats:

 - implicit hexadecimal RGB triple: e.g. ff7f3f
 - explicit RGB triple: e.g. rgb(255, 127, 63)
 - HSV (hue‑saturation‑value) triple: e.g. hsv(20, 75, 100)
     hue ∊ [0, 360] (degrees)
     saturation, value ∊ [0, 100] (percent)
     note: this is sometimes called HSB (hue‑saturation‑brightness)
 - HSL (hue‑saturation‑lightness) triple: e.g. hsl(20, 100, 62)
     hue ∊ [0, 360] (degrees)
     saturation, lightness ∊ [0, 100] (percent)

Example:

	liquidctl -m kraken set ring color fading 'hsv(0,80,100)' 'hsv(180,80,100)'

Related: #81 ("Kraken X62 and Smart Device v2 .Add Brightness
funcionality.")