Traverse all hid/hidraw queued reports and return the last one
As incoming HID reports are queued by the OS, HidapiDevice.read() would
return stale data to the drivers.  The drivers assume the reads aren't
queued because that's not necessary with the currently supported
devices.

This didn't affect the liquidctl CLI as the connection to HIDAPI
wouldn't last long enough for too much data to be queued and lost.  But
for downstream projects like GKraken this was a significant regression
from liquidctl 1.1.0.

To keep the behavior consistent between the HIDAPI and Libusb backends,
traverse all queued (incoming) reports until the last one is found, and
only return it.

Related: #87 ("1.3.2 regression: kraken_two.get_status() returns wrong
fan RPM value")
Reported-by: @leinardi