Add the need for setuptools (and pip) to manually install on Linux
A couple of users have reported issues installing liquidctl due to
missing setuptools.  In one of those instances, this happened with pip
on a Ubuntu-based distro.

setuptools should be a (pretty essential) dependency of pip, but the
latter wont be installed by default and, as mentioned above, the former
may be missing anyway.

Explicitly mention that they are needed, though leave pip as optional.