Clear enqueued HID reports on PyUsbHid/Windows
If PyUsbHid is used on a device bound to the Microsoft HID Driver, reads
from that device wont be as "raw" as previously thought and reports will
get queued.

Interally, libsusb uses ReadFile on such devices, which in turn reads
enqueued reports in a buffer.[1][2]

Similarly to HidapiDevice, implement clear_enqueued_reports by
successfully reading until there's nothing left in the buffer.

[1] https://github.com/libusb/libusb/blob/5c44ca955b3d7c41e6717514dc508f42faa2237b/libusb/os/windows_winusb.c#L3761
[2] http://janaxelson.com/hidfaq.htm

Fixes: #103 ("PyUsbHid.clear_enqueued_reports broken with Windows HIDs")