Release the device altogether on platforms other than Linux
On Windows, PyUsbDevice will likely be used with devices bound to the
WinUSB driver.  WinUSB only allows a single application at a time, and
this is likely to repeat with other drivers, and also possibly on mac
OS.

Make the more efficient behavior of only releasing the interface
exclusive to Linux, where it has been tested to work without issues (so
far).  On all other cases, release the device altogether.

Related: #101 ("Export liquidctl data to HWiNFO": "fix inability to
change cooler settings while the script is running (non HID only)")