Enable additional color channels
Channel IDs on HUE 2 devices are sent as a bit mask.

So if ring is 0b010 and logo is 0b100, 0b001 must be the 'external'
channel for accessories connected to the HUE 2 header—and this has been
confirmed in the captured traffic by @wamphlett— and a 'sync'
pseudo-channel can be enabled with 0b111.

The _STATIC_VALUE is currently assumed a number of LEDs to be used for
synchronizing the animations.  For now, set this value to 40 (HUE+
maximum) for both 'external' and 'sync'.

Related: 0ef5474a17d4 ("Send HUE 2 color channel as bit mask")