Increase log level to info for the invariant pump/logo accessories
Ideally I would like these to be listed on stdout with the remainder of
the output of initialize, but only when --verbose is passed.

However, there's no place yet in liquidctl where we set and use a common
'verbose' configuration.  So for now simply log it (stderr) with level
INFO.