Use a guard driver to require --debug for Kraken Z experiments
In preparation to merging the fourth generation drivers, restrict access
to the very early Z model driver to avoid accidental/unexpected
behavior.

This introduces the idea of a "guard driver", which is a part mock part
abstract (as in lacking implementation) stand-in for the real driver,
requiring `--debug` for access to the real driver.

Its purpose is to allow calls to initialize and status to succeed even
if the driver marked as 'future support' is disabled.  This prevents
breaking `liquidctl initialize all` or `liquidctl status` when there are
other fully supported devices in the system.