Remove the PyUsbHid device backend and deprecate --hid
PyUsbHid was originally introduced as a way to keep using PyUSB/libusb
for HIDs on all all platforms but Mac OS, where they were switched to
use HidapiDevice out of absolute necessity.

But over time the switch to Hidapi has proved to be beneficial on
the other major platforms as well:

- on Linux this allows the use of hidraw, with better compatibility with
  other programs and hwmon drivers;
- on Windows libusb is just a thin (and somewhat limited) wrapper over
  the HID handle, so switching to Hidapi allows more control and reduces
  the number of abstraction layers.

Additionally, PyUsbHid is essentially a reimplementation of the
libusb-based backend of hidapi.  In most cases where that is useful to
have, it is already provided by hidapi; sometimes, like in most Linux
distributions, even besides the other "true" HID implementation.

Remove the PyUsbHid backend.  This also obviously results in switching
Windows to use Hidapi.

Since on Linux we generally want to always want to use hidraw (if it is
available), also deprecate and ignore any value supplied to `--hid`.[1]

[1] As before, try to use the `hidraw` module and if it is not
available, either because it is the only supplied implementation
(ArchLinux) or because it was disabled at Hidapi build time, fallback to
the `hid` module.  The difference is that it is no longer possible to
override this behavior with `--hid hid` or `--hid usb`.