Use temporary file and rename pattern to make stores atomic
Atomic, but not durable, as this is *not* a requirement for the drivers:
they are prepared to handle missing data.

Weirdly damaged data could cause some issues, but as the values are
almost only a few bytes long, this is very unlikely (each value fits in
a single sector or block).