Include opaque prefix and profile length in set_cooling data
Trying to set fan speeds with liquidctl did not work and caused the
device to also refuse fan speed adjustments from iCue.[1][2]

While what we assume is the profile length was missing, that in unlikely
to be the only cause since the problem is not restricted to setups with
fan profiles.

It seems there is more important data to be sent.  Let us start with
the prefix, since it is by far the most unusual (remaining) part of the
of the captured packets.

Reported-by: @makk50
Related: #82 ("WIP: Add support for Corsair Platinum and PRO XT coolers")

[1] https://github.com/jonasmalacofilho/liquidctl/pull/82#issuecomment-644655719
[2] https://github.com/jonasmalacofilho/liquidctl/pull/82#issuecomment-644664868