Add driver for Gigabyte RGB Fusion 2.0 USB controllers
Work by @CaseySJ, thus credited as the author of this commit.

However, the driver and the tests had to be manually pulled in by
@jonasmalacofilho, as there were too many unrelated changes and a few
conflicts.

This was pulled-in from Casey's commit ee51be34da34 in PR #130, which
can be found at:

https://github.com/CaseySJ/liquidctl/commit/ee51be34da34d8719ae8bae968b151a8224565a4
https://github.com/jonasmalacofilho/liquidctl/commit/ee51be34da34d8719ae8bae968b151a8224565a4

A few minimal yet essential fixes (by @jonasmalacofilho) are also
included here, ensuring that the CLI and the unit tests remain
functional.  However, at this point, a few unit tests still fail.