Update the list of related projects of interest in 2020
__Removed:

KsenijaS/krakenx

Very interesting to me, and historically important to the development of
liquidctl.  But while it has been updated to use the liquidctl driver,
it does not support all features.  Additionally, a documentation issue
is now likely to cause (retroactive) issues with liquidctl.

jaksi/leviathan

The X31/X41/X61 coolers are loosing relevance and have quite a few quirks.
Additionally, leviathan is not a hwmon driver.

brkalmar/leviathan

Not a hwmon driver, and conflicts with liquidctl due to the working at
the USB layer (instead of the HID layer) and blocking Linux Hidraw.

__Added: OpenRGB