hydro_platinum: Remove component abstraction in set_color
This abstraction assumed each "component", meaning a fan or the pump
head, had the same number of LEDs.  It that held, the abstraction
provided a simple but convenient way to control each component
independently.

Unfortunately, the PRO XT devices made it obvious that the assumption
was incorrect: the pump head is very likely the same between Platinum
and PRO XT coolers, and it is now known to have 16 LEDs (on the PRO XT).

This makes the component abstraction a harder to (correctly) implement,
since the fans and the pump head have significantly different number of
LEDs.

Considering that it was also a very crude abstraction, a more direct
approach, where each component is identified by its name through
separate (pseudo) channels, is preferable to fixing the old one.

This will, however, require thorough testing, and is not something I am
ready to include (and thus stabilize) in 1.4.0.

For now, simply remove the broken abstraction and provide basic fully
independent or fully synchronized control over the LEDs.

Related: #149 ("H100i RGB PRO XT Lighting not responding to commands")
Related: #138 ("Add support for H150i PRO XT and H100i Platinum SE")