nzxt_epsu: don't inject report IDs
According to their HID descriptor, NZXT E-series PSUs do not use
numbered reports.  This is especially relevant when doing writes with
Hidapi, since it has to work with non-compliant devices and thus needs
to know the report ID, even if it's absent.

The problem appears to be aggravated on Windows, possibly because it
tries to fix the mess and ends up sending a report that is larger than
what the PSU expects/accepts.  But this is just conjecture...

Related: #166 ("NZXT E850 status now throws error with 1.4.0")