corsair_hid_psu: don't inject report IDs
According to their HID descriptor, Corsair HXi/RMi PSUs do not use
numbered reports.  This is especially relevant when doing writes with
Hidapi, since it has to work with non-compliant devices and thus needs
to know the report ID, even if it's absent.

The problem appears to be aggravated on Windows, possibly because it
tries to fix the mess and ends up sending a report that is larger than
what the PSU expects/accepts.

Related: #192 ("Unexpected error with Corsair HX1200i, 72 is not a valid
OCPMode")