nvidia: evga_pascal: remove unsafe=evga_pascal requirement
Provided SMBus remains opt-in (`unsafe=smbus` continues to be a
requirement), support for the EVGA GTX 1080 FTW should be as [un]safe as
other drivers in liquidctl, or other programs that can control the RGB
lighting in these cards.

- we only target a very specific model
- we can be very confident that we have the correct bus for that model
  (we comprehensively check that based on the PCI parent of the i2c bus,
  and there are unit tests for this)
- a bus is specific to a card, and not open to peripherals
- a specific address is used
- there is suitable test coverage for all writes, and reads are not know
  to cause harm

Tested on my card; matches the behavior of OpenRGB and, according to I²C
captures taken with NvAPISpy,[1] EVGA Precision XOC and EVGA LED Sync.

[1] https://gitlab.com/CalcProgrammer1/NvAPISpy