rgb_fusion2: linux: don't create driver instance for second usage
Hidapi 0.10 introduces support for multiple usage and usage pages to
hidraw, and may now return multiple handles for a single device
interface.

However, since previous versions of hidapi left the usage_page and usage
field unintialized, rather than set to a knwon invalid value (like 0),
we cannot positively check for the desired usage.

Instead, we check for undesired one, and assume that in all other cases
we either have the desired usage, or that on that system (platform,
hidapi version, and hidapi backend) a single handle is still returned
for the device.

Related: #259 ("Linux with hidapi-0.10.1 is showing RGB Fusion
controller twice")