asetek: release the entire device on Windows
Allow the simultaneous use of more than one instance of liquidctl (or
other liquidctl-based programs) on Windows.

While on Linux (and possibly on Mac and the BSDs) we can get by simply
releasing the interface, Windows requires the entire device to be
released for other programs to be able to access it.

Closes: #246 (yoda + liquidctl status on Windows)