docs: readme: temporarily remove link to 1.5.0 executable
Windows Defender has heuristically flagged it, based on machine
learning.[1]

This is most likely an error, since the file is built on AppVeyor, and
has not been modified afterwards (which we can tell because of
checksums).  Thus, I have submitted the file for manual analysis.

Still, trying to use it is very frustrating, since Defender keeps trying
to remove it.  So, for now, let us keep normal Windows users on 1.4.2.

[1] https://www.microsoft.com/en-us/wdsi/threats/malware-encyclopedia-description?Name=Trojan%3aWin32%2fFuerboos.D!cl