linux: use udev to load the i2c_dev module when needed
Removes extra/linux/modules-load.conf, and the need to install it.
The module is also only loaded if a supported i2c device is found.

That said, the entire i801_smbus (and, soon, i2c_piix4) counts as a
"supported i2c device", so i2c_dev will generally be loaded.

Additionally, a small downside to this approach is that, if the module
was loaded due to one of these rules, it is no longer possible to unload
it with modprobe -r.

Still, overall, this simplifies packaging and will probably be a net
positive change.