doc: readme: suggest python -m pytest instead of pytest
The latter may import some already installed liquidctl package, instead
of using the local sources, which is the desired behavior in the
context context of the particular example that we show.

While currently both python -m pytest and pytest work exactly the same
way (and will prefer the local sources over an already installed package),
that is only because we have a conftest.py file in the root, and, in the
future, we may want to remove it (for unrelated reasons).