commander_pro: fix tolerant handling of single channel name
The user is not required to use a specific channel name if there is only
one channel.

However, the implementation of the commander_pro driver had a bug that
resulted in a None channel id being used in that case. As a consequence
of that unexpected None, a TypeError was raised later in the code path.

$ liquidctl --match "lighting node code" set led1 color fixed 00ff00
ERROR: Unexpected error with Corsair Lighting Node Core (experimental):
TypeError: 'NoneType' object is not iterable

This specifically affects the Lighting Node Core, which has a single LED
channel, but fix the problem for a hypothetical single-fan-channel
device as well.

Reported-by: qkzk (on Discord)[^1]

[^1]: https://discord.com/channels/780568774964805672/801844203772313680/922402647527538698