smart_device: v1: move firmware and LED accessory to initialize()
According to our documented policy, this is not a breaking change:[^1]

> 4. The output from CLI commands is *not* guaranteed to be stable,
> 5. and neither are the items returned from `get_status()` or `initialize()`.

Additionally, some current users of that information already need to be
fixed and call/execute initialize in the first place.

Finally, this allows us to have get_status() return data from hwmon,
when possible, and where firmware and LED accessory information is not
available.

[^1]: docs/developer/process.md ("Stability and backward compatibility" section)