keyval: skip symlink test if on Windows with insufficient permissions
Creating symlinks on Windows is, by default, a privileged operation.
From the `os.symlink()` documentation:

> On newer versions of Windows 10, unprivileged accounts can create
> symlinks if Developer Mode is enabled. When Developer Mode is not
> available/enabled, the SeCreateSymbolicLinkPrivilege privilege is
> required, or the process must be run as an administrator.

While it is important to test that our keyval storage is consistent even
with pathological symlinks, we shouldn't break the test suite if we're
unable to run this test because the user/developer hasn't enabled
developer mode.

Fixes: #460