usb: make SUPPORTED_DEVICES private
This fixes a mistake from pre-1.0 liquidctl and gives us flexibility to
change the structure used to hold that data, if that turns out to be
beneficial in the future.

Additionally, the new class field is called _MATCHES, following what was
used in the later nvidia module.

As far as I can tell the only external user of this field is Coolero,
and I'll submit a patch shortly to prevent any breakage there.

Keeping a temporary alias during some deprecation period was considered,
but isn't trivial since this is a class not instance field.  As there's
only a single external user, it's easier to just fix it.

Finally, a typical Coolero installation generally bundles a specific
liquidctl version, and otherwise existing global setups shouldn't go
through the affected path (before an update) since it's in a test.