corsair_hid_psu: assert that the response matches the command sent
A rare and so far not well understood issue with these PSUs has been an
occasional exception being thrown from timedelta:

    OverflowError: Python int too large to convert to C int

In both cases for which we have debug infos, #463 and #192 (part), it
seems that the cause is an invalid response from the PSU, probably
because of a race with another program simultaneously accessing it.

(A third case has been reported on Discord, but so far no debug info has
been provided to us. Still, it seems likely that once again there's
another program (iCue) in use at the same time as liquidctl).

Check the response matches the command sent in _exec, and if that fails,
suggest that a possible conflict with another program might be the
cause.

Also fix a small issue with how the mock used for unit testing
automatically preloads reports for future impending reads, which up to
this point failed to consider the write bit (which, for the OCP mode,
can be both write=0 or read=1).

Fixes: #463
Related: #192