tree: remove "experimental" notes and suffixes
In the early days of the project it was important to communicate to our
users how well they could expect a device to work with liquidctl. This
was done, for new or still unstable devices, with an `e`/experimental
note (in the supported devices table) and an `(experimental)` suffix in
their description. However, it doesn't make sense to do that anymore:

- we have many more drivers, supported devices, and contributors, and so
  our capability to *accurately* gauge that, across all configurations,
  use cases, firmware versions, etc, has likely degraded;
- our development pace has reduced substantially, degrading the temporal
  resolution (and, thus, accuracy) of any such claim;
- and, on the other hand, *most drivers* either work well from the start
  or have quickly and easily been fixed.

Exceptionally, `(experimental)` suffixes have been kept for a few
untested EVGA and ASUS GPUs, since their support require the passing of
corresponding unsafe flags to be enabled.