extra: adopt .py extension in python scripts
The primary motivation is to allow them, and any doctests contained in
them, to be discovered by pytest. The tests fixed in 31998d666ded had
been broken for for 3 years!

The change also makes sense in more general terms, as (with the
exception of prometheus-liquidctl-exporter), the scripts are
cross-platform and therefore may run in contexts (Windows) without
support for #!.