yoda: make psutil optional
A request has been made for yoda to be packaged in Arch AUR liquidctl-git.
However, on Linux (and BSD) yoda has an extra dependency of psutil,
which is used to access the system sensors.

To avoid bloating the dependency list of liquidctl as a whole, while
still not requiring separate packages, make psutil into an optional
requirement.