commander_core: make these devices as still broken
To be honest, it may be the case that some of the issues have indeed
been solved with a8be6831eb51 (PR #687). However, there still seems to
be a long list of remaining issues, and I think we're starting to lose
track of what works and what doesn't.

This partially reverts commits 6426a06c8dfc.

Related: #520, #583, #598, #623, #705
Related: PR #687