test_cli: don't change the type of sys.argv when monkeypatching it
When setting up for the CLI tests, we inadvertently change the type of
sys.argv from list to tuple. That worked with OG docopt, but breaks
docopt-ng. More importantly, the Python docs say sys.argv is a _list_.

So even though docopt was able to handle it being a tuple, and docopt-ng
seems (from a cursory look at its source) to still handle it being a
string, let's use the correct type.

Fixes: #708