smbus: skip unreadable EEPROMs
With DDR5 and the spd5118 kernel driver on Linux 6.11, an `eeprom` may
be available in sysfs even if trying to read it returns ENXIO (No such
device or address).

While this is a bit weird a possibly not intended by the kernel, let's
just work around it for now: ignore an unreadable EEPROM and just log
the error at DEBUG level.

Related: #731