keyval: don't try to use umask and chmod on windows
Python 3.13 accepts file descriptors in `chmod` and disables our stubs
for umask and chmod calls. This caused `PermissionError` failures on all
tests that depend on keyval storage and, presumably, would break real
users as well.

As we don't really care about the sticky bit on Windows, force the use
of the stubs on that platform.